+++ plugins/moveit/src/index.ts
@@ -7,6 +7,7 @@ export type {
 } from "./types.js";
 export { MOVEIT_PROFILE, resolveMoveItTopics } from "./profile.js";
 export { buildMoveItSnapshot } from "./build-snapshot.js";
+export { MOVEIT_FAILURE_RECIPES, evaluateFailureRecipes } from "./failure-recipes.js";
 export { MOVEIT_PLUGIN_MANIFEST } from "./manifest.js";
 export { MoveItDock } from "./MoveItDock.js";
 export { useMoveItSnapshot } from "./hooks/useMoveItSnapshot.js";